feat: a means of faster execution for metalanguage "code"
This project has never been about speed, and generally there are much
better options if you want either efficient logic programming, efficient
logical inference, or efficient functional programming. However. It does
turn out that once type safety is taken for granted, it is possible to
"compile" metalanguage elements with no free terms into something that
runs relatively close to the speed of equivalent python code. In many
cases this doesn't really matter, but because simplification (as a means
of running code) always involves AST transformations that can get quite
costly, there are cases where it is worth doing this. (This recently
came up when implementing finite instances of Kratzer's ordering
semantics.)

This commit therefore adds a second version of simplification, that
works by doing this compilation. The compilation code itself was overall
pretty trivial to write, and most of the work in this commit went into
providing reasonable wrappers for working with compiled code.