fix: simplification issues with sets/tuples and identity
The compilation system was really prompted by some bugs in simplify that
caused it to take very inefficient routes for sets consisting only of
MetaTerms, because the handling wasn't sufficiently general. This commit
goes back and fixes the bug, for both set membership and equality. Now
that the compilation system exists, it turns out to be most convenient
to actually use it for these cases. The resulting simplification timings
still aren't *fast*, but they aren't completely appalling any more.

This change could probably be generalized a bit...