fix: stablize simplify behavior for various set identities
This commit ensures that simplification has uniform and expected
behavior, by default, where the identities {Set x: False} (the empty
set) and {Set x: True) (the domain set for x's type) are concerned. For
condition sets, this is all worryingly special-case-y. At the end of the
day, I have concluded that this is acceptable because the cases that
that this code is checking for are all natural targets for other
simplification heuristics, e.g. boolean simplification's goal (which it
often can't achieve) is to reduce expressions to a constant boolean
term. However, it is still worth considering how to generalize some of
this behavior.

The situation for compilation is different and a bit unfortunate. The
fact is that the domain set case simply can't be handled, because there
is no compilation target. There are some heuristics that do work for
specific special cases, but I think essentially the general identity
laws involve cases that require static, not dynamic, handling. So the
current idea is that set expressions could use a static simplify pass
before compiling. (Which of course may be true for other things as well,
though my testing shows that many of the simplification heuristics that
do work in execution are faster in execution, by a lot.) I probably will
continue to chip away at this.

Corresponding documentation updates, including more careful caveats
about using `ConditionSet` in compilation.