refactor: clean up BindingOp constructor
This commit brings BindingOp and subclasses in line with other
constructors, in matching the ast closely; the api that is being
replaced here is very old, and still existed primarily because it would
be a pain to change. It's still a bit overly flexible, at least at the
superclass level.