refactor: move quantifiers into their own module
These were always a bit questionable in the `boolean` module, and
planned future changes to these quantifiers will make this even less
obvious. It also just seems helpful to be able to put them in their own
module.

(no doc changes here, this will need to be in a separate commit.)