fix: prioritize type assignments differently
In `under_type_assignment`, in cases where an inferred type assignment
and a provided type assignment conflicted, the inferred one would be
prioritized. This apparently doesn't come up much (and it's a 9yo bug it
seems), but when it does, it leads to extremely messy situations. (In
the case I ran into it, a let-bound X1 was supposed to be mapped to a
freshened type, but an inferred value mapped that freshened type to X1,
leading to an X1 in a supposedlly completely freshened expression. This
then collided with X1 on type rebinding.)

This is complicated enough that I've written a targeted test that
expresses the exact problem.