feat: restricted quantification
This commit adds support for restrictors to all the quantifiers as well
as the iota operators.  I originally hesitated to add restricted
quantifiers because they're not the classical starting point, they're
more complicated, I couldn't decide how to implement restrictors, and
restriction can be implemented with logical operations (albeit
non-uniformly). However, in the context of recent work on evaluation,
it's extremely helpful (and arguably *less* complicated) to be able to
explicitly indicate finite quantifier domains. This commit reorients
everything so that unary quantification is a special case of restricted
quantification, where the (optional) restrictors are sets around the
variable's type. The classical starting point is therefore preserved,
with all of the other issues resolved. The commit is a somewhat major
refactor.

* The parser is refactored to generalize binding op header parsing.
* The main classes are now simply called `Exists`, `Forall`, `Iota`.
  There are strictly unary versions with the old names, but these are
  not used, just present for compatibility.
* Many generalizations to code that interacts with quantifier domains.
* New relatively extensive documentation for binding expressions
* test cases