fix: improvements to binding expression header parsing
This commit fixes a few things to do with binding expression header
parsing:

* Handle parentheses. After recent refactoring, anything parenthetical
  was just dropped. This now allows parentheses around a term sequence.
* Generalize some boilerplate code based on how the compositional DRT
  fragment implements multivars. Also fix some bugs.
* Implement multivar alpha conversion (which isn't actually used in this
  fragment).
* Simplify the compositional DRT fragment itself.

This commit fixes handling of parentheses generally in variable
sequences; previous refactors left things in a state where parenthetical
content was simply dropped. It also generalizes a number of things about
multivar parsing so that a subclass doesn't need to handle them. This is
mainly impactful on the compositional DRT fragment, which I've updated.