fix: implement __r*__ magic methods where possible
Previously, python code like `2 + te("2")` would raise, because there
was no __radd__ method defined (and so on for many operators). This
commit implements these, while working around some idiosyncracies of
python operator overloading. Unfortunately, as far as I can tell, these
idiosyncracies for the comparison operators (<, <=, >, >=) are just not
possible to work around. For all cases but these, the metalanguage
guarantees to build an expression with the order that the user
specified, even if the first element is not a TypedExpr. For these,
there are cases involving MetaTerms where it is just not possible in
principle to know what the user intended. (An extreme solution would be
to make MetaTerm not a subclass of Term; it's this subclass relation
that causes the issue.)

This commit also adds an exhaustive test for all the cases that work.