fix: handle vacuous `Exists` with a non-finite domain
Simplification of something like `Exists x_n : True` would fail, because
only the positive finite case was correctly handled. This led to
downstream crashes in multiply quantified cases, because it is expected
that any quantified expression with no free terms can simplify. This
commit also slightly tweaks handling of non-finite domains vs
enumerability, adds a corresponding test for this case, and adds
exceptions in the relevant uniqueness code so that the bug is more
apparent if further versions of it appear during random testing.