feat: allow constructing MetaTerm functions in parser
The parser was already handling dicts pretty much exactly, just not
wrapping them in any sort of TypedExpr. It turns out to be very, very
easy to let MetaTerm functions be constructed directly this way.
Ultimately it may be useful to figure out some sort of more elaborate
non-meta mapping-based function class, but for now this is quite handy.