fix: track locally free type variables correctly
This bug could result in losing track of a locally free type variable
during TypeEnv.merge in cases where there are term mappings. The bug was
mostly harmless, in that the type variable was still there and would
generally get picked up by a let wrapper, but it resulted in incorrect
type variable compaction.