fix: let-handling for simplify_all
The use of combinators in 3871bb08053 introduced a mixed let situation
that could result in uncompacted type vars during simplify_all. This is
analytically correct as far as I can see (though the situation is a bit
complicated). Previously it was rare to nonexistent that ? types would
be instantiated during simplify_all, but there is nothing in principle
that would prevent this or make it a bad idea, and combinator-based
implementations are exactly the sort of use case where this comes up.
Therefore, add let preservation/handling to simplify_all.

(This feels a bit ad hoc, which is why I've tried to carefully work
through the justification; it possibly means there are other cases where
let preservation is missing as well...)