fix: handle some non-simplifiable ExistsExact cases
When a vacuous ExistsExact with a type variable is embedded in another
ExistsExact that looks simplifiable (no free terms, etc), then the
simplification code would crash. This is essentially a false positive in
the coarse detection of whether a complex recursive ExistsExact should
be simplifiable that would be better to handle up front. However, this
is hard to do in a general way. This change essentially cancels the
simplification path once the problem is detected.