refactor: simplify BindingOp api
This was historically very, very flexible. d518752c6f5a cleaned up some
of this, bringing it in line with other classes. This commit removes
most of the remaining (fairly unnecessary) flexibility, and pushes some
of the remaining init code into BindingOpHeader. We now uniformly treat
`TypedTerm` as a container for the variable info, and provide no other
ways to specify it.

This commit also changes the type coercion implementation for Iota,
hopefully as a preface to some bigger simplifications here.