refactor: clean up LFun type inference
Update LFun to use the new approach to type adjustment. This code is so
much simpler it's spooky, but it is passing all tests and still seems to
work on some very complicated combinator code. Hopefully there's no
unpleasant surprises down the road.

One thing worth noting is that the old api provides mechanisms for the
local type adjustment code to directly manipulate the env, which LFun
used. As far as I can tell this is completely unnecessary on the new
approach. Also tbd is exactly how derivations are affected.