fix: improve a cryptic error message
When an incorrect value is provided to unification, it typically errors
out in the type_ranking check that is modified here, leading to a fairly
hard to understand KeyError unless you deeply understand what is going
on. This commit wraps the type ranking check in something much more
obvious to the user, as well as some help for the stale type on reload
case.