fix: two bugs related to let polymorphism
First: copying TypedTerms did not correctly preserve let-polymorphism in
the type env. This was masked by some accidental workarounds in type
inference code.

Second: once it was behaving as intended, this revealed that preserving
assignment constraints led to all sorts of problems when substituting
let-polymorphic functions, because their bound terms retained the wrong
assignment constraints. I'm not entirely sure if this is the right
solution, but it hasn't broken anything yet.