fix: some sort/normalization issues for meta elements
Constructing a MetaTerm from a structure that contains MetaTerms would
lead to recursive MetaTerms, because the wrapped domain element hashes
with the domain element. This was mostly benign, but causes crashes with
sorting code, because a wrapped domain element doesn't support the same
comparison operations as the domain element. This commit addresses this
by normalizing out the recursive MetaTerms. (Sort code is also tweaked,
but isn't the main fix.)