docs: prototype quarto refactor
This begins a refactor of the lambda notebook docs into the structure
used by quarto, in order to have better web integration. In the long
run, I would like to have all the documentation be available this way,
but right now, this only converts some basic intro material as well as
the demo (so that a commit of the executed notebook is no longer
needed). I've also refactored quite a bit of the demo.

Note: the structure here is pulled directly from
https://github.com/rawlins/svgling/tree/master/docs