fix: better handling for merging derivations via `derived`
If a `result` when calling `derived` had an existing derivation, the new
step would be merged in an extremely odd way. This cleans things up so
that `result.derivation` is treated as steps that happen following the
one `derived` is describing, which I think is how this case should
always be interpreted. It's possible this will have unintended side
effects that I have missed, as this is pretty core derivation code, but
the old behavior was so obviously wrong I think it just was not used
very much.