fix: a bug with compiled LFuns
When an LFun was embedded within an LFun and the outer context provided
a value for the outer variable, the collision was not handled correctly.
(In particular, it would get restored to the outer value before `inner`
was actually called.) This fix unfortunately involves an entire copy of
`context`, which seems like overkill, but there does not seem to be an
obvious alternative (that only uses `dict`s).