feat: metalanguage assignment syntax for operator definitions
This commit adds a proof of concept implementation of defining new
operators in a metalanguage magic. There's not a lot of guide rails or
documentation for this yet. This makes use of op_from_to and associated
code, so most of the changes here are just in parsing.

As part of this, I've refactored how assignment lines are parsed in the
first place. In order to handle the character `=` in operator names,
this needed something better than existing code, so all three assignment
modes are now more clearly disentangled with shared code handled
differently.