feat: TypedExpr facade
This implements a facade class for TypedExpr objects that exposes only
operator functions. An eventual goal is to use this directly in
metalanguage parsing, but unfortunately it can't be hashable and that
interferes with the current approach to parsing sets and metalanguage
map functions. Various cleanups to operator code in addition.