fix: handle order invariance in frozendict hash
Previously, a frozendict would hash differently depending on order,
mismatching dict itself. This mostly didn't matter for the minimal ways
it was used, but does matter in general. This commit also adds a proper
`__eq__`, relying on underlying dict `__eq__` (which is also order
invariant). Doing this revealed an unfortunate bug lurking in
`MetaTerm.__eq__`.