feat: a ternary conditional operator
This commit adds a ternary conditional operator modeled after
mathematical definitions by cases. This can be used a bit like a
programming "if...else..." construct, but since the metalanguage is
purely functional, its values are arbitrary typed elements (that must
have a consistent type). Functional languages like Haskell have
conditionals that work like this, but it's less like python's.

Right now it is not short-circuiting in simplification, but it may be
better if it is; details tbd.