feat: improved parsing features
This commit adds some better parsing tools based on monadic parsing
(/PEG parsing) and uses them specifically to parse assignment
expressions. The general aim here is to move towards fully parsing, in a
clean way, metalanguage expressions, though as of this commit,
metalanguage parsing is still delegated to the old parsers.

This commit adds a python >=3.10 dep, in order to use match expressions
for AST manipulation. We are just short of python 3.9's eol (Oct 2025)
at the time I write this commit message, so I think this should be ok.
(And some extant f-string code depends on later versions, though this
was unintentional and should be fixed.)