fix: don't drop derivations when building SingletonComposables
This was a regression in a790b6ad62366057, because `te` made a plain
copy (which excludes derivation) before the current commit when provided
a TypedExpr. The copy() call in factory is very old, so I'm not sure of
the motivation; this commit is a bit overengineered because it's unclear
to me whether what the right behavior is for calls to the factory vs
`te` vs the use of `te` in `SingletonComposable`. However, tests pass
and this does fix behavior in some provided notebooks.