refactor: rework Precondition
Currently, met_preconditions=True can happen in a ParseError either
because the code has no preconditions, or because there are non-trivial
preconditions that have been met. Join ideally will produce errors for
repetitions of the second type only, but this was hard to manage. This
commit tweaks Precondition so that all ParseErrors in its scope set a
new field that tracks its existence, and changes its scope so that any
following subsequence constructed using `+` to a precondition is in its
scope. I'm not 100% satisfied with this setup, but it works for now.