refactor: don't use exception handling for parser backtracking
The previous implementation extensively used `ParseError`s to signal
when the parser should backtrack. Conventional wisdom suggests that
while modern python try...except is essentially costless when there are
no exceptions raised, it cannot be assumed to be so when exceptions are
being raised and handled. This commit therefore moves to using return
values with an error set instead. It's slightly more error-prone in that
the erroring code is responsible for directly rolling back parser state
to the right position, and before this was implicitly handled by raising
vs returning. Otherwise, the logic is fairly similar in a lot of cases.

I have to say that on balance profiling does not reveal a major
improvement, so I'm not sure it is really true that exception
handling in modern python is so bad. However, there is a minor speedup
so I guess I'll keep this.