feat: experimental support for VSCode HTML+math rendering
Previously, I had not been able to get rendering for complex HTML output
that uses latex math code to work in VSCode. (Anything that rendered via
Latex or to some degree Markdown semi-worked.) This commit uses a recipe
from https://github.com/microsoft/vscode/issues/203725 to get this
working. Similar to the colab solution, this change injects KaTeX
autorender in a somewhat insane fashion. Unfortunately in this case, I
was unable to get a fully local solution to work (so this will probably
break without internet). The commit brute forces it on every output;
unlike colab, VSCode cells don't seem to be isolated so maybe there is a
way to install a hook, but I don't know (or care) enough about VSCode to
get that working right now.  The result is extremely inelegant but it
does seem to function.

It seems that one could write a vscode plugin building off of
markdown-math that integrated this change, but I don't have the js chops
to do this very quickly. As far as I can tell, no one has yet written
such a plugin, and the MS code is not often updated.